requirements.txt
absl-py==2.3.0
accelerate==1.7.0
addict==2.4.0
albucore==0.0.24
albumentations==2.0.7
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
annotated-types @ file:///croot/annotated-types_1709542908624/work
anyio==4.9.0
archspec @ file:///croot/archspec_1709217642129/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
bleach==6.2.0
boltons @ file:///croot/boltons_1737061692168/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
cfgv==3.4.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
clip==0.2.0
colorama==0.4.6
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
conda @ file:///croot/conda_1738168370816/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1736524596117/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1739298990519/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
contourpy==1.3.2
cryptography @ file:///croot/cryptography_1732129644106/work
cycler==0.12.1
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
defusedxml==0.7.1
distlib==0.3.9
distro @ file:///croot/distro_1714488253808/work
docopt==0.6.2
easydict==1.13
einops==0.8.1
elasticdeform==0.5.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
facexlib==0.3.0
fastjsonschema==2.21.1
filelock==3.13.1
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.58.0
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
humanfriendly==10.0
icecream==2.1.4
identify==2.6.12
idna @ file:///croot/idna_1714398848350/work
image-quality==1.2.7
imageio==2.37.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
iniconfig==2.1.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython==8.12.3
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
libsvm==3.23.0.4
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.8
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1698846045803/work
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1698845653285/work
menuinst @ file:///croot/menuinst_1738943416351/work
mistune==3.1.3
mpmath==1.3.0
narwhals==1.43.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.3
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
nvitop==1.5.1
onnxruntime==1.22.0
openai==1.84.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.2.3
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.2.1
pipreqs==0.5.0
piq==0.8.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
plotly==6.1.2
pluggy @ file:///croot/pluggy_1733169602837/work
pre_commit==4.2.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
protobuf==6.31.1
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1734736067156/work
pydantic_core @ file:///croot/pydantic-core_1734726052986/work
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
pyiqa==0.1.13
pyparsing==3.2.3
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytest==8.4.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq @ file:///croot/pyzmq_1734687138743/work
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
rich @ file:///croot/rich_1732638981168/work
rpds-py==0.25.1
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
setuptools==75.8.0
simsimd==6.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
soupsieve==2.7
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
stringzilla==3.12.5
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.24
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.15.2
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchvision==0.22.0+cu128
tornado @ file:///croot/tornado_1747918059467/work
tqdm @ file:///croot/tqdm_1738943501192/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.37.2
triton==3.3.0
truststore @ file:///croot/truststore_1736550121485/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
virtualenv==20.31.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
yapf==0.43.0
yarg==0.1.9
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard @ file:///croot/zstandard_1731356346222/work
